+++ core/changelog.py
@@ -97,6 +97,7 @@ class Version:
         for name, value in self.fields.items():
             embed.add_field(name=name, value=truncate(value, 1024), inline=False)
         embed.set_footer(text=f"Current version: v{self.bot.version}")
+
         embed.set_thumbnail(url=self.bot.user.display_avatar.url)
         return embed